// packages/json-joy/src/json-patch/codec/compact/types.ts
export type OPCODE_STARTS = OPCODE.starts | 'starts';
// ...
export type CompactStartsOp =
  | [opcode: OPCODE_STARTS, path: string | Path, value: string]
  | [opcode: OPCODE_STARTS, path: string | Path, value: string, ignore_case?: 1];